More work on ConsoleApp and FooSyncEngine guts
Starting to do work on FooSyncEngine things like RepositoryState which
need modifications to suit many-to-many syncing instead of the old
source-repository model.

Note that this change makes the old .FooSync_RepoState.dat files
incompatible: each repository now has a GUID associated with it, instead
of just the machine name like before, and this is stored as a new field
placed first in the file.