repository diff work (lots of changes)
One problem discovered: if two sync groups form partially disjoint sets,
when one group syncs, the RepositoryStateCollections for the shared
repositories will be updated, and then syncing the other group will use
one of those RSCs, and changes there will not be shown.

Possible solution: instead of going with the newest RSC, go with the
*oldest* RSC. However, if a sync happens that resolves conflicts, those
conflicts could show up again in syncing the other group. Need to
investigate this possibility.

Other possible solution: use each repo's RSC as the viewpoint when
looking at only that repo's files. Not sure if this is the right
approach, though, as it may make the Inspect algorithm impossibly
complicated.